Refactor application structure and improve error handling
- Disable error display in production environment for security.
- Update file inclusion paths to use absolute paths for better reliability.
- Replace deprecated SecurityUtils and Logger classes with new implementations.
- Remove obsolete check-php-compatibility.php script.
- Enhance response handling with a dedicated ResponseHandler class.
- Implement CSRF token generation and validation in SecurityUtils.
- Improve logging functionality to support both file and database logging.
- Ensure consistent formatting and error messages across the application.
